% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_download.R
\name{download_request}
\alias{download_request}
\title{Helper for downloading files from HydroShare}
\usage{
download_request(id = NULL, path = NULL, ...)
}
\arguments{
\item{id}{Resource ID}

\item{path}{Path in resource's contents}
}
\description{
Helper for downloading files from HydroShare
}
\keyword{internal}
