% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_null}
\alias{handle_null}
\title{Helper function for handling \code{NULL}
values returned from API calls.}
\usage{
handle_null(content, is_list = FALSE)
}
\arguments{
\item{content}{Content returned from element of
\link[httr:content]{httr::content}}

\item{is_list}{Is \code{content} a list?}
}
\description{
Helper function for handling \code{NULL}
values returned from API calls.
}
\keyword{internal}
