% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_params}
\alias{handle_params}
\title{Helper function for handling parameters
passed to \code{hs_resource}.}
\usage{
handle_params(...)
}
\arguments{
\item{...}{Parameters passed to any request.}
}
\description{
Algorithmically determine
what function is supposed to be called based on
the given parameters.
}
\keyword{internal}
