% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resource.R
\name{hs_access}
\alias{hs_access}
\title{Access Resource Permissions}
\usage{
hs_access(id, ...)
}
\arguments{
\item{id}{Resource ID}

\item{...}{Unused}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Access Resource Permissions
}
