% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hs_agent}
\alias{hs_agent}
\title{Helper for HSClientR's \link[httr:user_agent]{user_agent}.}
\usage{
hs_agent()
}
\description{
Helper for HSClientR's \link[httr:user_agent]{user_agent}.
}
\keyword{internal}
