% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hs_auth}
\alias{hs_auth}
\title{Helper for \link[httr:authenticate]{authentication}.}
\usage{
hs_auth(set_header = FALSE)
}
\value{
A \link[httr:Token-class]{Token} if \code{set_header = FALSE},
otherwise an access token set in global config
to be included in all requests.
}
\description{
Helper for \link[httr:authenticate]{authentication}.
}
\keyword{internal}
