% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsapi_GET.R
\name{hs_content_types}
\alias{hs_content_types}
\title{List Content Types}
\usage{
hs_content_types()
}
\value{
A character vector of content types.

As of 2/24/2021, the types returned will be:
\itemize{
\item \code{GenericLogicalFile}
\item \code{GeoRasterLogicalFile}
\item \code{NetCDFLogicalFile}
\item \code{GeoFeatureLogicalFile}
\item \code{RefTimeseriesLogicalFile}
\item \code{TimeSeriesLogicalFile}
\item \code{FileSetLogicalFile}
}
}
\description{
List Content Types
}
