% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resource.R
\name{hs_files}
\alias{hs_files}
\title{Get Resource Files}
\usage{
hs_files(id, page = NULL, count = NULL)
}
\arguments{
\item{id}{Resource ID}

\item{page}{Page of results to return}

\item{count}{Number of files to return per page}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get Resource Files
}
