% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resource.R
\name{hs_folder}
\alias{hs_folder}
\title{Get Resource Folder}
\usage{
hs_folder(id, pathname)
}
\arguments{
\item{id}{Resource ID}

\item{pathname}{Path to folder in resource's contents}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get Resource Folder
}
