% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resource.R
\name{hs_scimeta}
\alias{hs_scimeta}
\title{Get Resource Science Metadata}
\usage{
hs_scimeta(id)
}
\arguments{
\item{id}{Resource ID}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get Resource Science Metadata
}
