% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsapi_GET.R
\name{hs_search}
\alias{hs_search}
\title{Search HydroShare Resources using solr conventions.}
\usage{
hs_search(
  page,
  count,
  text,
  author,
  contributor,
  subject,
  abstract,
  resource_type,
  content_type,
  coverage_type,
  availability,
  created,
  modified,
  start_date,
  end_date,
  east,
  north,
  eastlimit,
  westlimit,
  northlimit,
  southlimit
)
}
\description{
Search HydroShare Resources using solr conventions.
}
