% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hs_session}
\alias{hs_session}
\title{Helper function to \link[httr:POST]{POST} authentication
and pass along session.}
\usage{
hs_session()
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\keyword{internal}
