% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_status.R
\name{hs_status}
\alias{hs_status}
\title{Get HydroShare API Status}
\usage{
hs_status()
}
\value{
Output indicating API status
}
\description{
This function calls upon the API base URL to return a status code 200.
If a code other than 200 is returned, the function will stop.
}
