% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resource.R
\name{hs_sysmeta}
\alias{hs_sysmeta}
\title{Get Resource System Metadata by ID}
\usage{
hs_sysmeta(id)
}
\arguments{
\item{id}{Alphanumeric HydroShare Resource ID}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get Resource System Metadata by ID
}
