% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_user.R
\name{hs_user}
\alias{hs_user}
\title{List Authenticated User Information}
\usage{
hs_user()
}
\value{
A \link[tibble:tibble-package]{tibble} of \emph{your} user details.
}
\description{
List Authenticated User Information
}
