% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsapi_GET.R
\name{hs_user_details}
\alias{hs_user_details}
\title{List User Details for Some User}
\usage{
hs_user_details(user_identifier)
}
\arguments{
\item{user_identifier}{HydroShare User Identifier}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Currently, lifecycle is experimental due to requiring
OAuth2 authentication. There is currently limited
documentation on OAuth2 authentication for HydroShare.
}
