% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hsapi_request}
\alias{hsapi_request}
\title{Helper function for:
\link[httr:GET]{GET}, \link[httr:POST]{POST},
\link[httr:PUT]{PUT}, and \link[httr:DELETE]{DELETE} requests.}
\usage{
hsapi_request(path, type, ...)
}
\arguments{
\item{path}{API Request Path}

\item{type}{Request Type (GET, POST, PUT, or DELETE)}

\item{...}{Parameters passed onto a \code{httr} request}
}
\description{
Helper function for:
\link[httr:GET]{GET}, \link[httr:POST]{POST},
\link[httr:PUT]{PUT}, and \link[httr:DELETE]{DELETE} requests.
}
\keyword{internal}
