% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSClient.R
\name{HSClient}
\alias{HSClient}
\title{R6 Class representing a HydroShare Client}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\emph{In development}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HSClient$new()}}
\item \href{#method-print}{\code{HSClient$print()}}
\item \href{#method-clone}{\code{HSClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a HydroShare Client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{HSClient} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print HSClient object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{HSClient} output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
