% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSClientR-package.R
\docType{package}
\name{HSClientR-package}
\alias{HSClientR-package}
\title{HydroShare HSClientR}
\description{
This package provides a \code{R6}-based client and set of functions
for accessing the \href{https://hydroshare.org}{HydroShare} API.
This client is originally modeled after the Python-equivalent
\href{https://github.com/hydroshare/hs_restclient}{hs_restclient}.
}
\author{
Justin Singh-Mohudpur \email{justin@justinsingh.me}
}
\keyword{package}
