% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hs_auth}
\alias{hs_auth}
\title{Helper for \link[httr:authenticate]{authentication}.}
\usage{
hs_auth(oauth = FALSE, client_id = NULL, client_secret = NULL)
}
\arguments{
\item{oauth}{Authenticate with OAuth2? \ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}}

\item{client_id}{OAuth2 Client ID (Key), required if \code{oauth = TRUE}}

\item{client_secret}{OAuth2 Client Secret (Secret), required if \code{oauth = TRUE}}
}
\description{
Helper for \link[httr:authenticate]{authentication}.
}
\keyword{internal}
