% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resource.R
\name{hs_resource}
\alias{hs_resource}
\title{List resources}
\usage{
hs_resource(...)
}
\arguments{
\item{...}{See \strong{details}.}
}
\value{
A \link[tibble:tibble-package]{tibble}. See \strong{details}.
}
\description{
This is the primary function for querying the
HydroShare API \verb{/hsapi/resource} endpoint. It will return
a \link[tibble:tibble-package]{tibble} corresponding to the
given parameters. Endpoint is chosen algorithmically based
on passed parameters.
}
\details{
Below is a sectioned list of all possible parameters
that \code{hs_resource} can parse, and the
corresponding HydroShare API endpoint(s)/functions used.

Parameters marked with an asterisk (*) are required for
that endpoint.
\subsection{Searching Resources (\code{hs_search})}{
\subsection{Endpoint:}{
\itemize{
\item \strong{/hsapi/resource/search}
}
}

\subsection{Parameters:}{\tabular{lll}{
   Parameter \tab Data/Object Type \tab Description \cr
   \code{page} \tab \code{integer} \tab A page number within the paginated result set. \cr
   \code{count} \tab \code{integer} \tab Number of results to return per page. \cr
   \code{creator} \tab \code{character} \tab The first author (name or email) \cr
   \code{author} \tab \verb{character vector} \tab List of authors (name or email) \cr
   \code{group} \tab \code{character} \tab A group name (requires \code{edit_permissions = TRUE}) \cr
   \code{user} \tab \code{character} \tab Viewable by user (name or email) \cr
   \code{owner} \tab \code{character} \tab Owned by user (name or email) \cr
   \code{from_date} \tab \code{Date} \tab To get a list of resources created on or after this date \cr
   \code{to_date} \tab \code{Date} \tab To get a list of resources created on or before this date \cr
   \code{subject} \tab \verb{character vector} \tab Comma separated list of subjects \cr
   \code{full_text_search} \tab \code{character} \tab Get a list of resources with this text \cr
   \code{edit_permission} \tab \code{logical} \tab Filter by edit permissions of user/group/owner? \cr
   \code{published} \tab \code{logical} \tab Filter by published resources? \cr
   \code{type} \tab \verb{character vector} \tab List of resources of the specified resource types. See below  possible values. \cr
   \code{coverage_type} \tab \code{character} \tab To get a list of resources that fall within the specified tial coverage boundary. Either: "box" or "point" \cr
   \code{coordinates} \tab \verb{character vector} \tab List of spatial coordinates in the form of (north, south, t, west) \cr
   \code{include_obselete} \tab \code{logical} \tab Include replaced resources? \cr
}


For available resource types, see \link{hs_types}.
}

}

\subsection{Getting Resource Information by ID (\code{hs_sysmeta})}{
\subsection{Endpoints:}{
\itemize{
\item \strong{/hsapi/resource/{id}/sysmeta}
}
}

\subsection{Parameters:}{\tabular{lll}{
   Parameters \tab Data/Object Type \tab Description \cr
   \code{id}* \tab \code{character} \tab Resource Alphanumeric ID \cr
}

}

\subsection{Return:}{\tabular{ll}{
   Column \tab Description \cr
   \code{resource_title} \tab Title of the resource \cr
   \code{resource_type} \tab Resource type \cr
   \code{resource_id} \tab ID of the resource \cr
   \code{abstract} \tab Resource abstract \cr
   \code{authors} \tab Authors of resource \cr
   \code{creator} \tab Creator of resource \cr
   \code{doi} \tab Resource DOI ID \cr
   \code{public} \tab Indicates if resource if public \cr
   \code{discoverable} \tab Indicates if resource is discoverable \cr
   \code{shareable} \tab Indicates if resource is shareable \cr
   \code{immutable} \tab Indicates if resource is immutable \cr
   \code{published} \tab Indicates if resource is published \cr
   \code{date_created} \tab Date resource was created \cr
   \code{date_last_updated} \tab Date resource was last updated \cr
   \code{bag_url} \tab URL to resource download, can be passed to \code{\link{download.file}} \cr
   \code{science_metadata_url} \tab URL to science metadata \cr
   \code{resource_map_url} \tab URL to resource map \cr
   \code{resource_url} \tab URL to resource \cr
   \code{content_types} \tab Resource content types, see \code{\link{hs_content_types}} \cr
}

}

}

\subsection{Getting Resource Access Permissions (\code{hs_access})}{
\subsection{Endpoint:}{
\itemize{
\item \strong{/hsapi/resource/{id}/access}
}
}

\subsection{Parameters:}{\tabular{lll}{
   Parameter \tab Data/Object Type \tab Description \cr
   \code{id}* \tab \code{character} \tab Resource Alphanumeric ID \cr
   \code{access}* \tab \code{logical} \tab Must be \code{TRUE} for access query \cr
}

}

\subsection{Return:}{\tabular{ll}{
   Column \tab Description \cr
   \code{id} \tab Resource Alphanumeric ID \cr
   \code{privilege} \tab User's privilege level \cr
   \code{user} \tab User to be granted privilege \cr
   \code{resource} \tab Resource to which privilege applies \cr
   \code{grantor} \tab Grantor of privilege \cr
}

}

}

\subsection{Getting Resource File List (\code{hs_files})}{
\subsection{Endpoint:}{
\itemize{
\item \strong{/hsapi/resource/{id}/files}
}
}

\subsection{Parameters:}{\tabular{lll}{
   Parameter \tab Data/Object Type \tab Description \cr
   \code{id}* \tab \code{character} \tab Resource Alphanumeric ID \cr
   \code{files}* \tab \code{logical} \tab Must be \code{TRUE} for files query \cr
   \code{page} \tab \code{integer} \tab Page of results to return \cr
   \code{count} \tab \code{integer} \tab Number of files to return per page \cr
}

}

\subsection{Return:}{\tabular{ll}{
   Column \tab Description \cr
   \code{file_name} \tab The filename, including the path \cr
   \code{url} \tab The url to download the file \cr
   \code{size} \tab The size of the file \cr
   \code{content_type} \tab The content type of the file \cr
   \code{logical_file_type} \tab The logical file type \cr
   \code{modified_time} \tab The last modified time of the file \cr
   \code{checksum} \tab The file's checksum \cr
}

}

}

\subsection{Getting Resource Folder(s) (\code{hs_folder})}{
\subsection{Endpoint:}{
\itemize{
\item \strong{/hsapi/resource/{id}/folders/{pathname}/}
}
}

\subsection{Parameters:}{\tabular{lll}{
   Parameter \tab Data/Object Type \tab Description \cr
   \code{id}* \tab \code{character} \tab Resource Alphanumeric ID \cr
   \code{pathname}* \tab \code{character} \tab Path to folder in resource's contents \cr
}

}

\subsection{Return:}{\tabular{ll}{
   Column \tab Description \cr
   \code{resource_id} \tab Resource Alphanumeric ID \cr
   \code{path} \tab Path to folder in resource's contents \cr
   \code{contents} \tab Nested \code{tibble} with files and folders within \code{path} \cr
}

}

}

\subsection{Getting Resource Science Metadata (\code{hs_scimeta})}{
\subsection{Endpoints:}{
\itemize{
\item \strong{/hsapi/resource/{id}/scimeta}
}
}

\subsection{Parameters:}{\tabular{lll}{
   Parameter \tab Data/Object Type \tab Description \cr
   \code{id}* \tab \code{character} \tab Resource Alphanumeric ID \cr
   \code{scimeta}* \tab \code{logical} \tab Must be \code{TRUE} for scimeta query \cr
}

}

\subsection{Return:}{\tabular{ll}{
   Column \tab Description \cr
   \code{id} \tab Resource Alphanumeric ID \cr
   \code{title} \tab Resource Title \cr
   \code{citation} \tab Resource Bibliography Citation \cr
   \code{creators} \tab Nested \code{tibble} of resource creator(s) \cr
   \code{sources} \tab Nested \code{tibble} of resource source(s) \cr
   \code{metadata} \tab Nested \code{tibble} with additional resource metadata \cr
}

}

}
}
