% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsapi_GET.R
\name{hs_resource}
\alias{hs_resource}
\title{List resources}
\usage{
hs_resource(
  page = NULL,
  count = NULL,
  creator = NULL,
  author = NULL,
  group = NULL,
  user = NULL,
  owner = NULL,
  from_date = NULL,
  to_date = NULL,
  subject = NULL,
  published = NULL,
  type = NULL,
  full_text_search = NULL,
  edit_permission = NULL,
  coverage_type = NULL,
  coordinates = NULL,
  include_obselete = NULL
)
}
\arguments{
\item{page}{(\code{integer}) A page number within the paginated result set.}

\item{count}{(\code{integer}) Number of results to return per page.}

\item{creator}{(\code{character}) The first author (name or email)}

\item{author}{(\verb{character vector}) List of authors (name or email)}

\item{group}{(\code{character}) A group name (requires \code{edit_permissions = TRUE})}

\item{user}{(\code{character}) Viewable by user (name or email)}

\item{owner}{(\code{character}) Owned by user (name or email)}

\item{from_date}{(\code{Date}) to get a list of resources created
on or after this date}

\item{to_date}{(\code{Date}) to get a list of resources created
on or before this date}

\item{subject}{(\verb{character vector}) Comma separated list of subjects}

\item{published}{(\code{logical}) filter by published resources?}

\item{type}{(\verb{character vector}) list of resources of the
specified resource types. \\
See details for possible values.}

\item{full_text_search}{(\code{character}) get a list of resources with this text}

\item{edit_permission}{(\code{logical}) filter by edit permissions
of user/group/owner?}

\item{coverage_type}{(\code{character}) to get a list of resources that fall
within the specified spatial coverage
boundary \\
Either: "box" or "point"}

\item{coordinates}{(\verb{character vector}) list of spatial coordinates in the
form of (north, south, east, west)}

\item{include_obselete}{(\code{logical}) Include replaced resources?}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns:
\itemize{
\item \code{resource_title} - Title of the resource
\item \code{resource_type} - Resource type
\item \code{resource_id} -  ID of the resource
\item \code{abstract} - Resource abstract
\item \code{authors} - Authors of resource
\item \code{creator} - Creator of resource
\item \code{doi} - Resource DOI ID
\item \code{public} - Indicates if resource if public
\item \code{discoverable} - Indicates if resource is discoverable
\item \code{shareable} - Indicates if resource is shareable
\item \code{immutable} - Indicates if resource is immutable
\item \code{published} - Indicates if resource is published
\item \code{date_created} - Date resource was created
\item \code{date_last_updated} - Date resource was last updated
\item \code{bag_url} - URL to resource download,
can be passed to \code{\link{download.file}}
\item \code{science_metadata_url} - URL to science metadata
\item \code{resource_map_url} - URL to resource map
\item \code{resource_url} - URL to resource
\item \code{content_types} - Resource content types,
see \code{\link{hs_content_types}}
}
}
\description{
List resources
}
\details{
Available resource types for \strong{type}:
\itemize{
\item \code{GenericResource}
\item \code{RasterResource}
\item \code{RefTimeSeriesResource}
\item \code{TimeSeriesResource}
\item \code{NetcdfResource}
\item \code{ModelProgramResource}
\item \code{ModelInstanceResource}
\item \code{ToolResource}
\item \code{SWATModelInstanceResource}
\item \code{GeographicFeatureResource}
\item \code{ScriptResource}
\item \code{CollectionResource}
\item \code{MODFLOWModelInstanceResource}
\item \code{CompositeResource}
}

Also returned by \code{\link{hs_types}}.
}
