% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resource.R
\name{hs_search}
\alias{hs_search}
\title{Search HydroShare Resources using solr conventions.}
\usage{
hs_search(page = NULL, count = NULL, text = NULL, author = NULL, ...)
}
\arguments{
\item{page}{(\code{integer}) A page number within the paginated result set.}

\item{count}{(\code{integer}) Number of results to return per page.}

\item{text}{(\code{character}) Search across all Resource Fields}

\item{author}{(\code{character}) Search by author}

\item{...}{Additional parameters passed to the HydroShare API.
See \strong{details}.}
}
\value{
A list with \code{.next}, \code{.prev}, and \code{results} attributes.
\code{results} returns a \link[tibble:tibble-package]{tibble}.
}
\description{
Search HydroShare Resources using solr conventions.
}
\details{
\subsection{Additional Parameters:}{\tabular{lll}{
   Parameter \tab Data/Object Type \tab Description \cr
   \code{contributor} \tab \verb{character vector} \tab Search by contributor(s) \cr
   \code{subject} \tab \verb{character vector} \tab Search within subject keywords \cr
   \code{abstract} \tab \code{character} \tab Search within the abstract \cr
   \code{resource_type} \tab \code{character} \tab Search by resource type \cr
   \code{content_type} \tab \verb{character vector} \tab Search by content type(s) \cr
   \code{coverage_type} \tab \verb{character vector} \tab Search by coverage type(s) (point, box, period) \cr
   \code{availability} \tab \verb{character vector} \tab Search by availabilities (discoverable, public, published) \cr
   \code{created} \tab \code{Date} \tab Search by created date \cr
   \code{modified} \tab \code{Date} \tab Search by modified date \cr
   \code{start_date} \tab \code{Date} \tab Search by start date \cr
   \code{end_date} \tab \code{Date} \tab Search by end date \cr
   \code{east} \tab \code{numeric} \tab Search by location or box center east longitude \cr
   \code{north} \tab \code{numeric} \tab Search by location or box center north latitude \cr
   \code{eastlimit} \tab \code{numeric} \tab Search by east limit longitude \cr
   \code{westlimit} \tab \code{numeric} \tab Search by west limit longitude \cr
   \code{northlimit} \tab \code{numeric} \tab Search by north limit latitude \cr
   \code{southlimit} \tab \code{numeric} \tab Search by south limit latitude \cr
}

}
}
