% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_user_details.R
\name{hs_user_details}
\alias{hs_user_details}
\title{List User Details for Some User}
\usage{
hs_user_details(user_identifier)
}
\arguments{
\item{user_identifier}{HydroShare User Identifier}
}
\value{
A \link[tibble:tibble-package]{tibble} of user details.
}
\description{
List User Details for Some User
}
