% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hsapi_request}
\alias{hsapi_request}
\title{Helper function for:
\link[httr:GET]{GET}, \link[httr:POST]{POST},
\link[httr:PUT]{PUT}, and \link[httr:DELETE]{DELETE} requests.}
\usage{
hsapi_request(path, type, query = NULL, auth = TRUE, oauth = TRUE)
}
\arguments{
\item{path}{API Request Path}

\item{type}{Request Type (GET, POST, PUT, or DELETE)}

\item{query}{API Query Parameters}

\item{auth}{Use basic HTTPS authentication}

\item{oauth}{Use OAuth2 authentication \ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}}
}
\description{
Helper function for:
\link[httr:GET]{GET}, \link[httr:POST]{POST},
\link[httr:PUT]{PUT}, and \link[httr:DELETE]{DELETE} requests.
}
\keyword{internal}
