% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSClient.R
\name{HSClient}
\alias{HSClient}
\title{R6 Class representing a HydroShare Client}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\emph{In development}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HSClient$new()}}
\item \href{#method-print}{\code{HSClient$print()}}
\item \href{#method-authenticate}{\code{HSClient$authenticate()}}
\item \href{#method-query}{\code{HSClient$query()}}
\item \href{#method-get_next}{\code{HSClient$get_next()}}
\item \href{#method-get_prev}{\code{HSClient$get_prev()}}
\item \href{#method-clone}{\code{HSClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a HydroShare Client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{HSClient} object.

\emph{Note: A message will be issued
since R6-object is still in development.}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print HSClient object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{HSClient} output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authenticate"></a>}}
\if{latex}{\out{\hypertarget{method-authenticate}{}}}
\subsection{Method \code{authenticate()}}{
Perform OAuth2 Authentication
with HydroShare.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$authenticate()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Note, the access token will be set
as a header automatically if \code{.token}
is \code{NULL}.
}

\subsection{Returns}{
A \link[httr:Token-class]{Token} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
Query/Search HydroShare
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$query(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Query parameters. See \link{hs_search} parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[tibble:tibble-package]{tibble}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_next"></a>}}
\if{latex}{\out{\hypertarget{method-get_next}{}}}
\subsection{Method \code{get_next()}}{
Get next resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$get_next()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{resource} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_prev"></a>}}
\if{latex}{\out{\hypertarget{method-get_prev}{}}}
\subsection{Method \code{get_prev()}}{
Get previous resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$get_prev()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{resource} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HSClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
